Delete old source code.
