--- src/Classes/AftershipAPIException.java
@@ -1,15 +1,0 @@
-package Classes;
-
-/**
- * Created by User on 10/6/14.
- */
-public class AftershipAPIException extends Exception{
-
-    public AftershipAPIException() {}
-
-    public AftershipAPIException(String message)
-    {
-        super(message);
-    }
-
-}
