--- src/Classes/query/TrackingResponse.java
@@ -1,12 +1,0 @@
-package Classes.query;
-
-import Classes.Tracking;
-
-import java.util.List;
-
-/**
- * Created by georgestoica on 30/3/16.
- */
-public interface TrackingResponse {
-    List<Tracking> getTrackingData();
-}
