--- src/Enums/StatusTag.java
@@ -1,8 +1,0 @@
-package Enums;
-
-/**
- * Created by User on 13/6/14.
- */
-public enum StatusTag {
-    Pending,InTransit,OutForDelivery,AttemptFail,Delivered,Exception,Expired,InfoReceived
-}
