--- src/Tests/CheckpointTest.java
@@ -1,18 +1,0 @@
-package Tests;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class CheckpointTest {
-
-    @Test
-    public void testPrueba() throws Exception {
-
-
-        // MyClass is tested
-      //  Checkpoint tester = new Checkpoint();
-    }
-
-
-}
